# Replace floating point color translations in Jpeg format