--- a/src/ImageSharp.Formats.Jpeg/JpegDecoderCore.cs
+++ b/src/ImageSharp.Formats.Jpeg/JpegDecoderCore.cs
@@ -607,10 +607,16 @@ namespace ImageSharp.Formats
             int ccb = cb - 128;
             int ccr = cr - 128;
 
-            byte r = (byte)(y + (1.402F * ccr)).Clamp(0, 255);
-            byte g = (byte)(y - (0.34414F * ccb) - (0.71414F * ccr)).Clamp(0, 255);
-            byte b = (byte)(y + (1.772F * ccb)).Clamp(0, 255);
+            // Speed up the algorithm by removing floating point calculation
+            // Scale by 1024, add .5F and truncate value. We use bit shifting to divide the result
+            int r0 = 1436 * ccr; // (1.402F * 1024) + .5F
+            int g0 = 352 * ccb; // (0.34414F * 1024) + .5F
+            int g1 = 731 * ccr; // (0.71414F  * 1024) + .5F
+            int b0 = 1815 * ccb; // (1.772F * 1024) + .5F
 
+            byte r = (byte)(y + (r0 >> 10)).Clamp(0, 255);
+            byte g = (byte)(y - (g0 >> 10) - (g1 >> 10)).Clamp(0, 255);
+            byte b = (byte)(y + (b0 >> 10)).Clamp(0, 255);
             packed.PackFromBytes(r, g, b, 255);
         }