Replaces the floating point calculations with scaled integers and bit-shifting to improve the performance of the jpeg format.

I've included some benchmarks for comparison.

Please have a look at it and test against any images that you know can cause issues. 

Still not happy with that decoder performance on my machine. We'll figure it out though I'm sure. 

### Jpeg Encode

**Before**

Method |      Mean |    StdDev | Scaled | Scaled-StdDev |    Gen 0 | Allocated |
---------------------- |---------- |---------- |------- |-------------- |--------- |---------- |
'System.Drawing Jpeg' | 4.9274 ms | 0.1229 ms |   1.00 |          0.00 | 271.8750 | 894.51 kB |
'ImageSharp Jpeg' | 9.2119 ms | 0.4067 ms |   1.87 |          0.09 |        - |  66.96 kB |

**After**

Method |      Mean |    StdDev | Scaled | Scaled-StdDev |    Gen 0 | Allocated |
---------------------- |---------- |---------- |------- |-------------- |--------- |---------- |
 'System.Drawing Jpeg' | 4.9281 ms | 0.0531 ms |   1.00 |          0.00 | 271.8750 | 894.51 kB |
'ImageSharp Jpeg' | 7.9708 ms | 0.2411 ms |   1.62 |          0.05 |        - |  66.96 kB |

### Jpeg Decode

**Before**

Method |       Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
---------------------- |----------- |---------- |------- |-------------- |---------- |
'System.Drawing Jpeg' | 11.1743 ms | 0.7354 ms |   1.00 |          0.00 | 260.74 kB |
'ImageSharp Jpeg' | 67.4747 ms | 1.6753 ms |   6.07 |          0.45 |   3.87 MB |

**After**

Method |       Mean |    StdDev | Scaled | Scaled-StdDev | Allocated |
---------------------- |----------- |---------- |------- |-------------- |---------- |
'System.Drawing Jpeg' | 11.3669 ms | 0.7688 ms |   1.00 |          0.00 | 260.74 kB |
'ImageSharp Jpeg' | 63.8410 ms | 0.4163 ms |   5.64 |          0.40 |   3.87 MB |