@antonfirsov 2017-01-11
@JimBobSquarePants I still hope it can be done better, maybe using integers is not the best. Want to play a bit with benchmarks first to elaborate the most optimal way. (around 11pm CET)
Is it ok if I push benchmark changes directly to the master?

I know the `(Encode|Decode)JpegMultiple` benchmarks demand more patience, but running them could provide more information on this :smile: