@JimBobSquarePants 2017-01-11
Using scaled integers with bit-shifting like this is a pretty standard way to optimize code like this. The original DCT code did this for example. 

Yeah push benchmarks to master if you like.