@antonfirsov 2017-01-11
Everything depends on the context. You need to be very cautious with float <-> int, float <-> byte, int <-> byte conversions, because they are quite expensive.

Regarding accuracy: Maybe we could adopt an image comparsion techniques from [MonoGame tests](https://github.com/MonoGame/MonoGame/blob/develop/Test/Framework/Visual/VisualTestFixtureBase.cs) in the future.