@JimBobSquarePants 2017-01-11
@antonfirsov True that... I wonder sometimes whether we are, to coin a phrase my mother uses trying to "make a silk purse out of a sows ear". basing our decoder on the Golang implementation. Yeah it works but it's so messy it feels like it has to be fundamentally architecturally flawed. 

I was very briefly looking at the codebase for Google's new jpeg decoder/encoder and it's a joy to read the code in comparison with Golangs. Everything is right where I'd expect it to be. 

https://github.com/google/guetzli

The way they color transform is interesting also. 
https://github.com/google/guetzli/blob/master/guetzli/color_transform.h#L211

I wonder whether we should try and build an alternative as an experiment?