@antonfirsov 2017-01-12
I'm also thinking about a full replacement sometimes, and this library looks really nice. I'm not happy with the current results, because on large images System.Drawing is still much more faster. Finding the reason is on my list, hope it's just multithreading.

However doing the "refactor to patterns" approach in the past 2 months, I realized that that if we do it the right way, in the end will be impossible to recognize the original golang mess. Your mom is right, evolution works like this! :smile: