@JimBobSquarePants 2017-01-12
I think multithreading will be a patch rather than a solution. We can beat System.Drawing with our Bmp format on a single thread so it should theoretically be possible to do the same (or at least come close with Jpeg) 

The refactoring work you've done so far is amazing but I don't want you to go so far down a path trying to refactor that codebase that it becomes a chore.