@antonfirsov 2017-01-12
Improved the benchmarks so they reflect the context of the problem better. Played with other algorithms, none of them can beat the simplest scalar integer approach :smile: 

`Vector<T>` really sucks. It can't even perform a vectorized dot product :cry: I wish I went to bed earlier instead.

I think you should merge your changes!

```
                                                  Method |          Mean |     StdDev | Scaled | Scaled-StdDev |
-------------------------------------------------------- |-------------- |----------- |------- |-------------- |
                             'Floating Point Conversion' |   537.3523 ns |  2.3920 ns |   1.00 |          0.00 |
                        'Simd Floating Point Conversion' |   767.8779 ns |  1.9016 ns |   1.43 |          0.01 |
                             'Scaled Integer Conversion' |   497.1790 ns |  1.9141 ns |   0.93 |          0.01 |
               'Scaled Integer Conversion + Vector<int>' |   604.9561 ns |  1.8976 ns |   1.13 |          0.01 |
 'Scaled Integer Conversion + Vector<int> + Dot Product' | 7,053.9099 ns | 31.2488 ns |  13.13 |          0.08 |
                         'Scaled Integer LUT Conversion' |   715.6920 ns |  1.9436 ns |   1.33 |          0.01 |
```