@antonfirsov 2017-01-12
Tested your changes the `*JpegMultiple` benchmarks. The optimization factor is significant. On small images now we are faster than System.Drawing!

[ORIGINAL](Jpeg-ScaledInteger-Optimization-ORIGINAL.md)

[OPTIMIZED](https://gist.github.com/antonfirsov/01743d39d53cb68f7071d935308c8ab7)

Regarding the deep refactor of the Jpeg codec:
I discourage now to start a full port-to-C#-from-scratch approach: at the current state it can bring much more "chore" work than a full clenup of our current codebase. The most (performance) critical part of the codec is the scan encoder/decoder part (DCT algorithms zigzagging etc.). It's now optimized for the CLR, covered with tests, and in the case of the decoder, moved to it's own class (JpegScanDecoder). It's prepared for further optimizations like parallelization.

I'd rather suggest to follow a mixed approach: Maybe we can port an existing library like guetzli (or parts of it), and replace the scan part with our already CLR-optimized one. 
I spent large amounts of time digging for ideas to speedup the DCT part. Have not found anything that could be used with CLR. Maybe there is something "around" the SOS part, I've missed, but unlikely.

However, I will continue my research. Are there any specific features you want to see in our codecs other than ICC profiles and progressive encoding?