@JimBobSquarePants 2017-01-12
@anton That's great news on the performance front!

I would certainly also prefer to keep our current work if we can continue to refactor the code to increase performance and reduce overheads and pinch bits from other libraries. We do seem to have a pressing issue with the jpeg encoder though. 

Download and take a look at https://github.com/bleroy/core-imaging-playground and run the LoadResizeSave method in Program.cs (Don't bother running the benchmarks)

Our jpeg output is awful compared to the others even if I set the supsampling to 4:4:4 and the quality to 100% (Oddly 4:4:4 is worse than 4:2:0 which it simply shouldn't be)

I'm thinking of digging out the old encoder I wrote in the [jpeg branch](https://github.com/JimBobSquarePants/ImageSharp/tree/2fb98f9642167e8556638d89b79e9c453950b45a/src/ImageProcessorCore/Formats/Jpg/New) updating the code and having a look at the output quality there and, if the quality is better, attempt to discern what the differences are.