@antonfirsov 2017-01-13
Wanted to play with gif octrees and other perf related stuff, but it seems I gonna have a jpeg weekend :D