@antonfirsov 2017-01-15
Okkay I have to confess! 

The quality drop has been introduced to the encoder with the `Block8x8F` refactor :(
Don't know yet, if it's an essential weakness of the floating point FDCT implementation or it's a mistake somewhere around it's application.

I'm going to investigate this issue. If it's not possible to use a floating point SIMD alogorithm without a quality drop, I will refactor the code to use integer blocks again. 

I will also think about the guetzli port.