@JimBobSquarePants 2017-01-15
J'accuse! 😝 

Thanks for investigating this and I'm glad you found a cause. I'll leave it in your capable hands for now and I won't merge this in either. (We'll see whether we need it first)