@antonfirsov 2017-01-16
@JimBobSquarePants I made a rounding bug in `Block8x8F.UnZigDivRound()`. Fixed and pushed it. I see no quality difference compared to the old integer-based version now. However, `System.Drawing`  quality is still better.

Check out `JpegTests.LoadResizeSave()` if you want to compare our current results to bleroy-s ones. Added 2 of it's test images.