@JimBobSquarePants 2017-01-16
Good work. I've just pushed a change that changes our subsampling rules to match those of Photoshop 4:4:4 default falling to 4:2:0 at <=50% quality. That improves the output quality but increases the size. Gonna experiment with the quality + this PR's changes now.