@JimBobSquarePants 2017-01-16
I'm finding the output is actually more pleasing to my eye (I tiny bit sharper for bleroys images; closer to ImageMagick) so once the CI checks pass I'm merging it.