@JimBobSquarePants 2017-01-16
I can actually pretty much mimic ImageMagick using the Lanczos3Resampler, our filesize is larger though as  we use standard Huffman tables rather than optimized. 

http://www.impulseadventure.com/photo/optimized-jpeg.html